requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
google-generativeai
python-docx==1.1.0
PyMuPDF==1.24.1
fpdf==1.7.2
sqlalchemy